requirements.txt
connexion[uvicorn,swagger-ui,flask]
flask
sqlalchemy
python-dotenv
psycopg2-binary
pytest